requirements.txt
flask
torch==2.1.0
pyod
pygod
pandas
transformers==4.38.2
langchain==0.1.0
fastapi
uvicorn
networkx
kafka-python
prometheus_client
streamlit==1.32.0
Pillow==10.0.0
